Move ScheduledTask to a different module and properly implement Ord for it
